Make servlet context listener be called from @WebListener
