+++ flow-server/src/main/java/com/vaadin/flow/server/startup/DevModeInitializer.java
@@ -17,8 +17,6 @@ package com.vaadin.flow.server.startup;
 
 import javax.servlet.ServletContainerInitializer;
 import javax.servlet.ServletContext;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRegistration;
 import javax.servlet.annotation.HandlesTypes;
@@ -29,7 +27,6 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,7 +35,6 @@ import com.vaadin.flow.component.WebComponentExporter;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.dependency.NpmPackage;
 import com.vaadin.flow.function.DeploymentConfiguration;
-import com.vaadin.flow.function.SerializableRunnable;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.server.DevModeHandler;
 import com.vaadin.flow.server.VaadinServlet;
@@ -108,31 +104,6 @@ public class DevModeInitializer
         }
     }
 
-    private static class StopDevMode
-            implements ServletContextListener, Serializable {
-
-        private final SerializableRunnable stopCallback;
-
-        private static final AtomicInteger SERVLET_CONTEXTS = new AtomicInteger();
-
-        private StopDevMode(SerializableRunnable stopCallback) {
-            this.stopCallback = stopCallback;
-        }
-
-        @Override
-        public void contextInitialized(ServletContextEvent sce) {
-            SERVLET_CONTEXTS.incrementAndGet();
-        }
-
-        @Override
-        public void contextDestroyed(ServletContextEvent sce) {
-            if (SERVLET_CONTEXTS.decrementAndGet() == 0) {
-                stopCallback.run();
-            }
-        }
-
-    }
-
     @Override
     public void onStartup(Set<Class<?>> classes, ServletContext context)
             throws ServletException {
@@ -196,9 +167,7 @@ public class DevModeInitializer
         context.setAttribute(VisitedClasses.class.getName(),
                 new VisitedClasses(visitedClassNames));
 
-        DevModeHandler handler = DevModeHandler.start(config,
-                builder.npmFolder);
-        context.addListener(new StopDevMode(() -> handler.stop()));
+        DevModeHandler.start(config, builder.npmFolder);
     }
 
     private static Logger log() {
