+++ flow-server/src/main/java/com/vaadin/flow/server/startup/ServletContextListeners.java
@@ -17,7 +17,8 @@ public class ServletContextListeners implements ServletContextListener {
      * The servlet must be deployed before websocket support is added to it.
      */
     private ServletContextListener[] listeners = new ServletContextListener[] {
-            new ServletDeployer(), new JSR356WebsocketInitializer() };
+            new ServletDeployer(), new JSR356WebsocketInitializer(),
+            new StopDevMode() };
 
     @Override
     public void contextInitialized(ServletContextEvent sce) {
